Redesign - part 5:
Changed the signature of the abstract Module class: instead of an array of expected connections, it is now a varargs-parameter. When a module is not expected to have direct connections to other modules (like a BIOS), the varargs is simply empty opposed to the previous array being passed as 'null'.