Redesign - part 7:
Removed the Modules class which held all Modules inside the Emulator class (extended an ArrayList) and fetched a specific Module by iterating over the Lits and comparing a `type:String` of the Module being searched for. 
Replaced it with a private `Map<Module.Type, Module>` attribute inside the Emulator class.